Retargeted project from Universal to iPhone only - the UI has been designed for a small screen and I doubt  anybody would ever use the remote on an iPad. Removed defaultPort as an instance attribute - it is a value common to all connections and was modelled wrongly. Boxee scan infra structure in place: scanner singleton created along with the scanning delegate; added CocoaAsyncSocket pod to the project (it will be used to perform the UDP broadcast).

+++ BoxeeRemote/Controllers/ConnectionSettingsViewController.m
@@ -46,7 +46,7 @@ static const NSInteger TXT_PASSWORD_TAG = 3;
     BoxeeConnection *lastConnection = [[BoxeeConnectionManager sharedManager] lastSuccessfulConnection];
     if (lastConnection) {
         self.txtBoxeeHost.text = lastConnection.hostname;
-        self.txtBoxeePort.text = [NSString stringWithFormat:@"%d", lastConnection.port];
+        self.txtBoxeePort.text = [NSString stringWithFormat:@"%ld", (long)lastConnection.port];
         self.txtBoxeePassword.text = lastConnection.password;
     }
 }
